Moved tests into their own directory
